Move catalog mutator to its own module
