+++ crates/catalog/src/lib.rs
@@ -1,4 +1,5 @@
 //! Session catalog definitions and interface to metastore.
 pub mod client;
 pub mod errors;
+pub mod mutator;
 pub mod session_catalog;
