+++ crates/sqlexec/src/planner/physical_plan/drop_database.rs
@@ -1,4 +1,4 @@
-use catalog::session_catalog::CatalogMutator;
+use catalog::mutator::CatalogMutator;
 use datafusion::arrow::datatypes::Schema;
 use datafusion::arrow::record_batch::RecordBatch;
 use datafusion::error::{DataFusionError, Result as DataFusionResult};
